+++ src/fmalloc.cpp
@@ -5,4 +5,5 @@
 #include "fmalloc_altrep.inc"
 #include "fmalloc_ops.inc"
 #include "fmalloc_tensor.inc"
+#include "fmalloc_alp.inc"
 #include "fmalloc_api.inc"